moving multi objective to its package
